+++ TaskHouseUnitTests/FakeRepositorys/FakeEmployerRepository.cs
@@ -1,59 +1,63 @@
-using TaskHouseApi.Model;
+using Xunit;
 using TaskHouseApi.Controllers;
-using System;
-using System.Collections.Concurrent;
+using Microsoft.AspNetCore.Mvc;
+using TaskHouseApi.Model;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Linq;
 using System.Threading.Tasks;
-using System.Data;
+using TaskHouseApi.Model.ServiceModel;
 using TaskHouseApi.Persistence.Repositories.Interfaces;
 
-namespace TaskHouseUnitTests
+namespace TaskHouseUnitTests.FakeRepositorys
 {
-    public class FakeWorkerRepository : FakeUserRepository<Worker>, IWorkerRepository
+    public class FakeEmployerRepository : FakeUserRepository<Employer>, IEmployerRepository
     {
-        public FakeWorkerRepository()
-       : base(new List<Worker>()
-       {
-                new Worker()
+        public FakeEmployerRepository()
+        : base(new List<Employer>()
+            {
+                new Employer()
                 {
-                    Id = 4,
-                    Username = "w1",
+                    Id = 1,
+                    Username = "1234",
                     Password = "+z490sXHo5u0qsSaxbBqEk9KsJtGqNhD8I8mVBdDJls=", //1234
                     Email = "test@test.com",
                     FirstName = "Bob1",
                     LastName = "Bobsen1",
                     Salt = "upYKQSsrlub5JAID61/6pA==",
-                    Discriminator = "Worker"
+                    Discriminator = "Employer"
+
                 },
-                new Worker()
+                new Employer()
                 {
-                    Id = 5,
-                    Username = "w2",
+                    Id = 2,
+                    Username = "root",
                     Password = "gCekoOdivUyOosjAz2wP9X+8TEEpe/LWJekDuvXV8bQ=", //root
                     Email = "test@test.com",
                     FirstName = "Bob2",
                     LastName = "Bobsen2",
                     Salt = "Ci1Zm+9HbvPCvVpBLcSFug==",
-                    Discriminator = "Worker"
+                    Discriminator = "Employer"
                 },
-                new Worker()
+                new Employer()
                 {
-                    Id = 6,
-                    Username = "w3",
+                    Id = 3,
+                    Username = "hej",
                     Password = "dpvq1pIWkY9SudflCKrW6tqCItErcBljM1GhNPWlUmg=", //hej
                     Email = "test@test.com",
                     FirstName = "Bob3",
                     LastName = "Bobsen3",
                     Salt = "U+cUJhQU56X+OCiGF9hb1g==",
-                    Discriminator = "Worker"
+                    Discriminator = "Employer"
                 }
-       }
-       )
+            }
+        )
         { }
         // public override IEnumerable<User> RetrieveAll()
         // {
-        //     return list.Where(w => w.Discriminator == nameof(Worker));
+        //     return list.Where(w => w.Discriminator == nameof(Employer));
         // }
     }
 }
+
+
