+++ TaskHouseUnitTests/UnitTests/AuthServiceUnitTest.cs
@@ -4,8 +4,9 @@ using TaskHouseApi.Persistence.Repositories.Interfaces;
 using TaskHouseApi.Persistence.UnitOfWork;
 using TaskHouseApi.Service;
 using Xunit;
+using TaskHouseUnitTests.FakeRepositorys;
 
-namespace TaskHouseUnitTests
+namespace TaskHouseUnitTests.UnitTests
 {
     public class AuthServiceUnitTest
     {
